rename stops
