+++ src/api/departures-grouped/schema.ts
