+++ src/service/impl/departures-grouped/journey-gql/departure-group.graphql
