+++ src/service/impl/departures-grouped/journey-gql/departure-time.graphql-gen.ts
