+++ src/service/impl/departures-grouped/utils/__tests__/fixture-venue-layer-favorites.json
